requirements.txt
<<<<<<< HEAD
# Core Libraries
numpy
pandas
scikit-learn
matplotlib
umap-learn
hdbscan

# AWS Integration
boto3

# LLM and API Access
openai

# Document Parsing
PyMuPDF  # for PDF parsing (imported as fitz)
python-docx  # for DOCX parsing
markdown  # for Markdown parsing

# CLI Interface
typer[all]  # includes Typer + rich output + click under the hood

# JSON Schema Validation (if you extend metadata validation)
jsonschema

# Progress Bars
tqdm

# Optional: Dev/Testing
pytest
pytest-mock
python-dotenv  # if you want to externalize secrets later

# Pin versions for stability (optional but recommended for production)
# numpy==1.26.0
# pandas==2.2.1
# scikit-learn==1.4.1.post1
# matplotlib==3.8.4
# umap-learn==0.5.5
# hdbscan==0.8.33
# boto3==1.34.82
# openai==1.17.0
# PyMuPDF==1.24.2
# python-docx==1.1.0
# markdown==3.6
# typer==0.12.3
# jsonschema==4.21.1
# tqdm==4.66.2
=======
# Core Libraries
numpy
pandas
scikit-learn
matplotlib
umap-learn
hdbscan

# AWS Integration
boto3

# LLM and API Access
openai

# Document Parsing
PyMuPDF  # for PDF parsing (imported as fitz)
python-docx  # for DOCX parsing
markdown  # for Markdown parsing

# CLI Interface
typer[all]  # includes Typer + rich output + click under the hood

# JSON Schema Validation (if you extend metadata validation)
jsonschema

# Progress Bars
tqdm

# Optional: Dev/Testing
pytest
python-dotenv  # if you want to externalize secrets later

# Pin versions for stability (optional but recommended for production)
# numpy==1.26.0
# pandas==2.2.1
# scikit-learn==1.4.1.post1
# matplotlib==3.8.4
# umap-learn==0.5.5
# hdbscan==0.8.33
# boto3==1.34.82
# openai==1.17.0
# PyMuPDF==1.24.2
# python-docx==1.1.0
# markdown==3.6
# typer==0.12.3
# jsonschema==4.21.1
# tqdm==4.66.2
>>>>>>> c21c4fb7d974b7cb11b9a77ae226a0dff2fad39c
